userservice create

+++ src/main/java/org/example/repository/TaskRepository.java
@@ -3,5 +3,5 @@ package org.example.repository;
 import org.example.entity.TaskEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskRepository extends JpaRepository<TaskEntity,Integer> {
+public interface TaskRepository extends JpaRepository<TaskEntity,Long> {
 }
